fix: set color values to unsigned shorts.
The reason there are no unsigned chars is, because chars represent characters, we want numbers. I truly do not care about tricks, so even though char can go from a number and back I straight up refuse to use it like that.
There is also an int_8 type but we can not assume all compilers have it. For saving ourselves a headache I have decided not to use it.

+++ include/engine/public/util/color.h
@@ -2,9 +2,9 @@
 
 class Color {
 public:
-    int r;
-    int g;
-    int b;
-    int a;
-    Color(int r, int g, int b, int a);
+    unsigned short r;
+    unsigned short g;
+    unsigned short b;
+    unsigned short a;
+    Color(unsigned short r, unsigned short g, unsigned short b, unsigned short a);
 };
