+++ src/public/util/color.cpp
@@ -1,4 +1,8 @@
 #include <engine/public/util/color.h>
 
-Color::Color(const int r, const int g, const int b, const int a): r(r), g(g), b(b), a(a) {
+Color::Color(
+    const unsigned short r,
+    const unsigned short g,
+    const unsigned short b,
+    const unsigned short a): r(r), g(g), b(b), a(a) {
 }
